FastAPI/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
pyautogui>=0.9.54
Pillow>=10.0.0
mss>=9.0.0
